+++ c#/Compiler/Compiler/obj/Debug/net10.0/Compiler.AssemblyInfo.cs
@@ -14,7 +14,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("Compiler")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+5f625ef9e3a96bc52ae4dab437af2fdbb447473a")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+cfb366e61834d74b08a0bdacdeebf088a1d1f52f")]
 [assembly: System.Reflection.AssemblyProductAttribute("Compiler")]
 [assembly: System.Reflection.AssemblyTitleAttribute("Compiler")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]